cfwebsocket addition, updates to network protocols and enhancement to findCategory function
Added cfwebsocket.json with basic info, code examples and helpful links.

Updated how findCategory looped thru the categories so it would look at
tags and functions last, which gives it the ability to find a tag or
function in any of the other categories first.

Updates to tags inside network protocol