Update arrayfind.json
Updated the description of using arrayFind with a closure / callback function. Previously, it was unclear what arrayFind() would return in the event that an array contains the item being searched for more than once. Updated to indicate that the index returned is the index of the FIRST found item in the array.