listAppend returns the concatenated list
This doc is a bit confusing when you are new to CF because unlike every other CF append function (array, struct) it isn't atomic, even when the member syntax is used. I added this to the description and the code examples to highlight this. listPrepend's doc clearly shows the proper use. Which is how I figured out what was wrong in my own code.