ArrayReverse example using different syntax in ACF2018
Not quite sure how to document this. ArrayReverse does not exist in ColdFusion but you can get the same result by using the array slice syntax `[::-1]` adding in ACF 2018. Have added an example  of doing it in ACF2018, but not sure how you'd normally handle this as it's an equivalent.