Fix terminateOnTimeout docs for cfexecute
I'm not sure about the listed version numbers, but it looks like there's some confusion.

The ACF docs history section only references a change made in ACF6.1 (adding the `variable` attribute), with no mention of initial addition of the method.

Likewise, the Lucee docs don't mention when the method was added.

For that reason, I've removed the version numbers referenced here. Hopefully the feature is old enough that it doesn't matter any more. :)