Issue #1447 - Finished note in description
The cftransaction page has a sidenote in its description that does not complete the sentence. To resolve this issue, I researched where this sidenote could have surfaced. It was found in the CFMX7 manual according to Newsgroup_User in the Adobe Support Community responsing to this very topic. I finished the sentence as part of the reference.

Q&A - [https://community.adobe.com/t5/coldfusion-discussions/cftransaction-amp-multiple-data-sources/m-p/844868](https://community.adobe.com/t5/coldfusion-discussions/cftransaction-amp-multiple-data-sources/m-p/844868)


> Newsgroup_User
> LEGEND , May 25, 2006
> This has been copied and pasted straight from the CFMX7 manual -
> "Within a transaction block, you can write queries to more than one
> database, but you must commit or roll back a transaction to one database
> before writing a query to another."