Added values specifications
Added values to difficulty param and specified height and width to **ONLY** work with pixels without the "px"

Furthermore, added related link to *imagecreatecaptcha()* that is the equivalent function in Adobe ColdFusion. As for minimum version of Lucee, my tests concluded that it would be 5.0.0.252 or later. I haven't test 4.x versions outside of the Lucee Documentation page. The option for "4.LATEST" does not seem to execute this function.